% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_flagging.R
\name{analyte_flagging}
\alias{analyte_flagging}
\title{Analyte Flagging "MUST BE RUN FIRST ON RAW OUTPUT AND BEFORE TIME FLAGGING"}
\usage{
analyte_flagging(
  path,
  onoff = " ",
  trantype = " ",
  h2smdl = NULL,
  h2shs = NULL,
  ch4mdl = NULL,
  ch4hs = NULL,
  benmdl = NULL,
  benhs = NULL,
  tolmdl = NULL,
  tolhs = NULL,
  xypmdl = NULL,
  xyphs = NULL
)
}
\arguments{
\item{path}{path to the raw MEM output from GMAP}

\item{onoff}{Must be "on" or "off". Whether the recording were done onsite "on" i.e. on facility property - within fenceline, or offsite "off" i.e. off facility property - outside of facility fenceline}

\item{trantype}{Must be "MA" or "ST". Whether the recorded transect was a "MA", mapping i.e GMAP vehicle moving or "ST", stationary i.e. GMAP vehicle stationary}

\item{h2smdl}{The Hydrogen sulfide Minimum Detection Limit in ppb}

\item{h2shs}{The Hydrogen sulfide high span in ppb}

\item{ch4mdl}{The Methane Minimum Detection Limit in ppm}

\item{ch4hs}{The Methane high span in ppm}

\item{benmdl}{The Benzene Minimum Detection Limit in ppb}

\item{benhs}{The Benzene high span in ppb}

\item{tolmdl}{The Toluene Minimum Detection Limit in ppb}

\item{tolhs}{The Toluene high span in ppb}

\item{xypmdl}{The p-Xylene Minimum Detection Limit in ppb}

\item{xyphs}{The p-Xylene high span in ppb}
}
\value{
a dataframe of analytes flagged according to concentrations of threshholds defined by the Minimum Detection Limit, and high span
}
\description{
Analyte Flagging "MUST BE RUN FIRST ON RAW OUTPUT AND BEFORE TIME FLAGGING"
}
