\name{buildGmapDbSNPIndex}
\alias{buildGmapDbSNPIndex}
\title{Download and Integrate dbSNP into GMAP/GSNAP}
\usage{buildGmapDbSNPIndex(version)}
\description{Downloads a version of dbSNP and integrates it into an installation of GMAP/GSNAP for SNP-tolerant alignment}
\details{.. content for details ..}
\value{0}
\author{Cory Barr}
\arguments{\item{version}{version of dbSNP to obtain}}
