\name{buildGmapIndex}
\alias{buildGmapIndex}
\title{Create Gmap Reference Genome Index}
\usage{buildGmapIndex(genome)}
\description{Gsnap index of hg19 (with haplotypes and unassembled regions excluded)}
\details{a reference genome into gmap/gsnap by downloading a
genome and constructing an IIT. Currently only hg19 is supported.}
\value{0}
\author{Cory Barr}
\arguments{\item{genome}{}}
