\name{consolidateBAMFiles}
\alias{consolidateBAMFiles}
\title{Sort and merge a set of BAM files}
\usage{consolidateBAMFiles(bam_files, outfile)}
\description{Helper utility to sort and merge a set of BAM files}
\details{<details>}
\value{a string indicating the name of the outfile}
\author{Cory Barr - Regular}
\arguments{\item{bam_files}{A character vector of the BAM files to be sorted}
\item{outfile}{The name of the resulting sorted and merged BAM
file (optional)}}
