% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_csv_data.R
\name{csv_output_data}
\alias{csv_output_data}
\title{output_csv_data}
\usage{
csv_output_data(x, newname = " ", onoff = " ", trantype = " ")
}
\arguments{
\item{x}{the final flagged GMAP data frames}

\item{newname}{name of the campaign or specific user provided name to identify the final files}

\item{onoff}{are the data collected onsite or offsite (inside vs outside the fenceline)}

\item{trantype}{The transect type either "MA" mapping, or "ST" stationary}
}
\value{
returns a csv in the active working directory of the final flagged results
}
\description{
output_csv_data
}
