% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_csv_max.R
\name{csv_output_max}
\alias{csv_output_max}
\title{csv_output_max}
\usage{
csv_output_max(input, newname = " ", onoff = " ", trantype = " ")
}
\arguments{
\item{input}{dataframe of flagged GMAP data}

\item{newname}{name of the campaign or specific user provided name to identify the final files}

\item{onoff}{are the data collected onsite or offsite (inside vs outside the fenceline)}

\item{trantype}{The transect type either "MA" mapping, or "ST" stationary}
}
\description{
csv_output_max
}
