\name{directory}
\alias{directory}
\title{Get the Path to the Location on Disk from a gmapR Class}
\description{
  Many objects in gmapR represent data stored on disk. The
  \code{directory} accessor will return this directory.
}
\usage{directory(x)}
\arguments{
  \item{x}{A GmapGenome or GmapSnps object}
}
\value{a character vector}
