\name{GmapSnps-class}
\Rdversion{1.1}
\docType{class}
\alias{GmapSnps-class}
\alias{directory,GmapSnps-method}

\title{Class \code{"GmapSnps"}}
\description{
  This class represents a set of SNPs (single nucleotide polymorphisms)
  for use with GMAP and GSNAP (typically for SNP-tolerant alignment.)
}
\section{Objects from the Class}{
  ##TODO: doc these args
Objects can be created by calls of the form \code{GmapSnps(snps,
  directory, name, create)}.
}
\section{Methods}{
  \describe{
    \item{directory}{\code{signature(x = "GmapSnps")}: ... }
  }
}
\author{
  Michael Lawrence
}
\keyword{classes}
