\name{GsnapParam}
\alias{GsnapParam}
\title{GsnapParam}
\usage{GsnapParam(unique_only=FALSE, max_mismatches, suboptimal_levels=0,
    mode="standard", use_snps, npaths=if (unique_only) 1L else 100L,
    quiet_if_excessive=unique_only, nofails=unique_only, format="bam",
    split_output=!unique_only, novelsplicing=FALSE, splicing,
    nthreads=1, part, batch="2", ...)}
