\name{installGmap}
\alias{installGmap}
\title{Download and Install GMAP}
\usage{installGmap(install_dir)}
\description{Downloads the GMAP source code, builds it, then installs it to a specified directory}
\value{0}
\author{Cory Barr}
\arguments{\item{install_dir}{directory where GMAP will be installed}}
