% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxes_hist.R
\name{maxes_hist}
\alias{maxes_hist}
\title{max histograms}
\usage{
maxes_hist(
  df,
  analyte = c("H2S(ppb)", "CH4(ppm)", "BEN(ppb)", "TOL(ppb)", "XYP(ppb)")
)
}
\arguments{
\item{df}{data frame of flagged GMAP data}

\item{analyte}{vector of analytes to be included in the analysis}
}
\value{
returns a plot of the maxes for each transect for the vector of analytes provided
}
\description{
max histograms
}
