% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_hist_virdis.R
\name{maxes_hist_virdis}
\alias{maxes_hist_virdis}
\title{maxes histogram virdis color}
\usage{
maxes_hist_virdis(
  df,
  analyte = c("H2S(ppb)", "CH4(ppm)", "BEN(ppb)", "TOL(ppb)", "XYP(ppb)")
)
}
\arguments{
\item{df}{dataframe of processed GMAP dataa}

\item{analyte}{the analyte(s) of interest in the max visualization}
}
\value{
histogram plot of the maxes for each transect using the virdis color pallete
}
\description{
maxes histogram virdis color
}
