% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_id.R
\name{NA_id}
\alias{NA_id}
\title{#identification of NA in the data (NAs indicate that the system was switched to dry air)}
\usage{
NA_id(path = " ", trantype = "", analyte = c("BEN", "TOL", "XYP"))
}
\arguments{
\item{path}{path to the raw output of the GMAP file}

\item{trantype}{The transect type either "MA" mapping, or "ST" stationary}

\item{analyte}{The analytes to check for NAs}
}
\value{
returns data frame with NAs removed
}
\description{
#identification of NA in the data (NAs indicate that the system was switched to dry air)
}
