% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_flagging.R
\name{time_flagging}
\alias{time_flagging}
\title{Time flagging:}
\usage{
time_flagging(
  x,
  timestart = " ",
  timestop = " ",
  timeqt = " ",
  analyte = c("H2Sflag", "CH4flag", "BENflag", "TOLflag", "XYPflag")
)
}
\arguments{
\item{x}{The output of analyte_flagging.Mus be a dataframe}

\item{timestart}{The time at which to begin applying the desired flag must be in the form of "YYYY-mm-dd HH:MM:SS"}

\item{timestop}{The time at which to end applying the time flag must be in the form of "YYYY-mm-dd HH:MM:SS"}

\item{timeqt}{The specific flag that the user wants to apply across the subset of times}

\item{analyte}{Analytes which to apply the specific flag}
}
\value{
Returns a dataframe with flagged data
}
\description{
Flagging analytes over specific temporal subsets: May need to be run multiple times if there are multiple subsets of times a user wants to apply a selected flag to multiple temporal subsets.
}
