% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Title}
\usage{
time_series(
  x,
  analyte = c("H2S(ppb)", "CH4(ppm)", "BEN(ppb)", "TOL(ppb)", "XYP(ppb)")
)
}
\arguments{
\item{analyte}{}
}
\description{
Title
}
