% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_figure.R, R/time_series_plot.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time series plot}
\usage{
time_series_plot(
  x,
  analyte = c("H2S(ppb)", "CH4(ppm)", "BEN(ppb)", "TOL(ppb)", "XYP(ppb)")
)

time_series_plot(
  x,
  analyte = c("H2S(ppb)", "CH4(ppm)", "BEN(ppb)", "TOL(ppb)", "XYP(ppb)")
)
}
\arguments{
\item{x}{flagged analyte data frame}

\item{analyte}{vector of analytes}
}
\value{
returns plots for each data/ transect combo in the plotting window

returns plots for each data/ transect combo in the plotting window
}
\description{
Time series plot

Time series plot
}
