% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_time_min_max.R
\name{tran_time_min_max}
\alias{tran_time_min_max}
\title{Maximum and minimum time for each transect from a GMAP campaign}
\usage{
tran_time_min_max(input)
}
\arguments{
\item{input}{Dataframe output from either analyte flagging or time flagging}
}
\value{
Returns a dataframe with the name of each transect and the beginning and ending times
}
\description{
Can be used to identify beginning and ending times fro each transect. Would be useful in conjunction with time
flagging if specific flags need to be applied to entire transects.
}
